# Use mul! in Diagonal*Matrix