--- a/stdlib/LinearAlgebra/src/diagonal.jl
+++ b/stdlib/LinearAlgebra/src/diagonal.jl
@@ -226,6 +226,7 @@ end
 function (*)(A::AbstractMatrix, D::Diagonal)
     Tdest = promote_op(*, eltype(A), eltype(D.diag))
     if eltype(A) <: Number && eltype(D) <: Number
+        Base.require_one_based_indexing(A)
         nA, nD = size(A, 2), length(D.diag)
         if nA != nD
             throw(DimensionMismatch("second dimension of A, $nA, does not match the first of D, $nD"))
@@ -237,6 +238,7 @@ end
 function (*)(D::Diagonal, A::AbstractMatrix)
     Tdest = promote_op(*, eltype(A), eltype(D.diag))
     if eltype(A) <: Number && eltype(D) <: Number
+        Base.require_one_based_indexing(A)
         nA, nD = size(A, 1), length(D.diag)
         if nA != nD
             throw(DimensionMismatch("second dimension of D, $nD, does not match the first of B, $nA"))