--- a/stdlib/LinearAlgebra/src/diagonal.jl
+++ b/stdlib/LinearAlgebra/src/diagonal.jl
@@ -143,6 +143,7 @@ iszero(D::Diagonal) = all(iszero, D.diag)
 zero(D::Diagonal) = Diagonal(zero.(D.diag))
 isone(D::Diagonal) = all(isone, D.diag)
 one(D::Diagonal) = Diagonal(one.(D.diag))
+oneunit(D::Diagonal) = Diagonal(oneunit.(D.diag))
 isdiag(D::Diagonal) = all(isdiag, D.diag)
 isdiag(D::Diagonal{<:Number}) = true
 istriu(D::Diagonal, k::Integer=0) = k <= 0 || iszero(D.diag) ? true : false