@MasonProtter 2021-09-20
Interesting finding! Is there any reason to not implement this with broadcasting though? Here are the numbers I'm seeing
```julia
julia> using BenchmarkTools: @btime

julia> using LinearAlgebra: Diagonal, mul!, similar, promote_op, transpose

julia> foreach((10, 50, 100, 500, 1_000, 5_000, 10_000)) do N
           M = N + rand(-5:5)
           A = rand(N, M)
           Dr = Diagonal(float.(1:M))

           println((;N, M))
           println("Current algorithm")
           @btime $A * $Dr
           println("Proposed algorithm")
           @btime $A *̂ $Dr
           println("Broadcasting")
           @btime $A .* transpose($Dr.diag)
           println()
       end
(N = 10, M = 14)
Current algorithm
  212.190 ns (3 allocations: 1.31 KiB)
Proposed algorithm
  192.421 ns (3 allocations: 1.31 KiB)
Broadcasting
  141.649 ns (1 allocation: 1.22 KiB)

(N = 50, M = 48)
Current algorithm
  2.612 μs (4 allocations: 18.92 KiB)
Proposed algorithm
  1.439 μs (4 allocations: 18.92 KiB)
Broadcasting
  1.346 μs (2 allocations: 18.83 KiB)

(N = 100, M = 99)
Current algorithm
  8.623 μs (4 allocations: 77.55 KiB)
Proposed algorithm
  4.386 μs (4 allocations: 77.55 KiB)
Broadcasting
  4.286 μs (2 allocations: 77.45 KiB)

(N = 500, M = 495)
Current algorithm
  212.060 μs (4 allocations: 1.89 MiB)
Proposed algorithm
  124.020 μs (4 allocations: 1.89 MiB)
Broadcasting
  126.190 μs (2 allocations: 1.89 MiB)

(N = 1000, M = 999)
Current algorithm
  1.336 ms (4 allocations: 7.62 MiB)
Proposed algorithm
  799.228 μs (4 allocations: 7.62 MiB)
Broadcasting
  801.688 μs (2 allocations: 7.62 MiB)

(N = 5000, M = 5001)
Current algorithm
  64.819 ms (4 allocations: 190.77 MiB)
Proposed algorithm
  43.458 ms (4 allocations: 190.77 MiB)
Broadcasting
  43.699 ms (2 allocations: 190.77 MiB)

(N = 10000, M = 9997)
Current algorithm
  266.747 ms (4 allocations: 762.71 MiB)
Proposed algorithm
  160.720 ms (4 allocations: 762.71 MiB)
Broadcasting
  159.481 ms (2 allocations: 762.71 MiB)
```
It seems that just using broadcast is simpler and just as fast or faster.


_____

```julia
julia> versioninfo()
Julia Version 1.6.1
Commit 6aaedecc44 (2021-04-23 05:59 UTC)
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  CPU: AMD Ryzen Threadripper 2990WX 32-Core Processor
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-11.0.1 (ORCJIT, znver1)
Environment:
  JULIA_NUM_THREADS = 26
```