@MasonProtter 2021-09-20
Normally, matrix multiplication involves a mixture of broadcast **and** reduction, but involving a `Diagonal` explicitly removes the reduction part and you're left with just an elementwise operation.