@N5N3 2021-09-21
In fact `mul!()`/`rmul!()`/`lmul!()` for Diagonal is implemented via broadcast.

There should be no runtime performance difference between 3-arg `mul!()` and manual broadcast, as the `alpha` and `beta` are propagated as const.