@N5N3 2021-09-21
The difference holds only for small martix. It might comes from (I didn't verify):
1. You manual broadcast use lazy `transpose`, while `mul!` use `premutedims`, which is create a new `Matrix`.
2. `mul!(z,x,y::Diagnal)` calculate z .= x .* permutedims(y.diag) .+ false. Maybe the `.+` was not omited by the optimizer?