@MasonProtter 2021-09-21
Hm. That eager `permutedims` Should probably be revisited.