@N5N3 2021-09-21
`@code_typed` shows that their loop body have almost the same IR.
except `mul!` with some `dead code`:
```julia
│     %308 = Base.mul_float(%304, %307)::Float64
│     %309 = Base.copysign_float(0.0, %308)::Float64
│     %310 = Base.ifelse(true, %308, %309)::Float64
│     %311 = Base.add_float(1.0, %310)::Float64
│     %312 = Base.ifelse(false, %311, %310)::Float64 
```
LLVM should be able to omit them.
Besides that `mul!` seems to perform more checks before the loop kernal, maybe this also contributes to the speed gap for small matrix.