@jishnub 2021-09-21
The difference for small matrices seems to be entirely due to `permutedims` being an array reshape that allocates:

```julia
julia> A = ones(10, 14); DR = Diagonal(ones(size(A, 2)));

julia> @btime permutedims($DR.diag);
  98.516 ns (2 allocations: 96 bytes)

julia> @btime transpose($DR.diag);
  2.674 ns (0 allocations: 0 bytes)

julia> @btime $A * $DR;
  439.354 ns (3 allocations: 1.31 KiB)

julia> @btime $A .* transpose($DR.diag);
  344.735 ns (1 allocation: 1.22 KiB)
```