@jishnub 2021-09-21
Would it make sense to improve `mul!` in a separate PR? Changing `StrideMatrix` to `AbstractMatrix` might need to be tested thoroughly.