@N5N3 2021-09-21
Both seems OK to me.
At present, in test.jl, `lmul!`, `rmul!` and `mul!` appears  11, 10, and 13 times, seperately.
While `*` appears 170 times.
Even if we assume that half of them are the reference for comparison, `*` is tested much more thoroughly.
Without this PR, we can say that `lmul!` and `rmul!` are well tested via `*`.
After this PR, things turns to be `mul!` is tested via `*`, while `lmul!` and `rmul!` not.
So it might be ok to do this change in this PR?