@dkarrasch 2021-09-21
> Hm. That eager `permutedims` Should probably be revisited. Seems like a waste of allocations that was done in the past because we didn’t handle `struct`s containing mutable data well.

No, we use `permutedims` because `transpose` operates recursively, but all we want is to "reshape" the vector into a row vector for broadcasting. If there is another (lazy) way of telling broadcasting that the vector should be interpreted as 1xn, then we can go with that. Otherwise, products with matrices of matrices will just go wrong.