@MasonProtter 2021-09-21
> No, we use permutedims because transpose operates recursively, but all we want is to "reshape" the vector into a row vector for broadcasting.

That shouldn’t be a problem if the method is restricted to operating on `Diagonal{T, Vector{T}} where T <: Number` though right? 

I guess if someone made some wacky number type that could be `transpose`’d that could be a problem?