@N5N3 2021-09-22
Since this PR accelerate `(*)(A::AbstractMatrix, D::Diagonal)`, I think at least an optimized `(*)(A::AbstractTriangular, D::Diagonal)` could be implemented via `(*)(A::AbstractMatrix, D::Diagonal)` to avoid data copy. 
For example:
```julia
@inline function _copy_diag!(data, diag)
    for i in 1:length(diag)
        data[i,i] = D.diag[i]
    end
    data
end
(*)(A::UpperTriangular, D::Diagonal) = UpperTriangular(A.data * D)
(*)(A::UnitUpperTriangular, D::Diagonal) = UpperTriangular(_copy_diag!(A.data * D, D.diag))
```